split the config and tas
